+++ lib/k8s/client/mint/request.ex
@@ -1,14 +1,19 @@
-defmodule K8s.Client.Mint.Request.WebSocket do
+defmodule K8s.Client.Mint.Request do
   @moduledoc """
-  Represents a WebSocket connection state.
+  Maintains the state of a HTTP or Websocket request.
 
   ### Fields
 
   - `:caller_ref` - Monitor reference of the calling process.
   - `:stream_to` - StreamTo requests only: The process expecting response parts sent to.
   - `:pool` - The PID of the pool so we can checkin after the last part is sent.
-  - `:websocket` - The websocket state (`Mint.WebSocket.t()`).
+  - `:websocket` - For WebSocket requests: The websocket state (`Mint.WebSocket.t()`).
   """
+
+  alias K8s.Client.Mint.ConnectionRegistry
+
+  @data_types [:data, :stdout, :stderr, :error]
+  @type request_types :: :sync | :stream_to | :stream
 
   @type t :: %__MODULE__{
           caller_ref: reference(),
@@ -17,10 +22,39 @@ defmodule K8s.Client.Mint.Request.WebSocket do
           websocket: Mint.WebSocket.t() | nil
         }
 
-  defstruct [:caller, :caller_ref, :stream_to, :pool, :websocket]
+  defstruct [:caller_ref, :stream_to, :pool, :websocket]
 
   @spec new(keyword()) :: t()
-  def new(fields \\ []), do: struct!(__MODULE__, fields)
+  def new(fields), do: struct!(__MODULE__, fields)
+
+  @spec put_response(t(), :done | {atom(), any()}) :: :pop | {:stop, t()} | {t(), t()}
+  def put_response(request, :done) do
+    send(request.stream_to, {:done, true})
+    Process.demonitor(request.caller_ref)
+    ConnectionRegistry.checkin(%{pool: request.pool, adapter: self()})
+    :pop
+  end
+
+  def put_response(request, {:close, data}) do
+    send(request.stream_to, {:close, data})
+    {:stop, request}
+  end
+
+  def put_response(request, {type, new_data})
+      when type in @data_types do
+    send(request.stream_to, {type, new_data})
+    {request, request}
+  end
+
+  def put_response(request, {type, value}) do
+    send(request.stream_to, {type, value})
+    {request, request}
+  end
+
+  @spec map_response({:done, reference()} | {atom(), reference(), any}) ::
+          {:done | {atom(), any}, reference()}
+  def map_response({:done, ref}), do: {:done, ref}
+  def map_response({type, ref, value}), do: {{type, value}, ref}
 
   @spec map_frame({:binary, binary} | {:close, any, any}) ::
           {:close, {integer(), binary()}}
